Projekt refactored
