+++ API/build.gradle.kts
@@ -1,4 +1,7 @@
 plugins {
+    id("org.springframework.boot") version "2.6.5"
+    id("io.spring.dependency-management") version "1.0.11.RELEASE"
+    id("com.google.cloud.tools.jib") version "3.2.1"
     java
 }
 
@@ -8,5 +11,7 @@ repositories {
 
 dependencies {
 
-    implementation("org.springframework.boot:spring-boot-starter-validation")
+
+    // https://mvnrepository.com/artifact/javax.validation/validation-api
+    implementation("javax.validation:validation-api:2.0.1.Final")
 }
