+++ API/src/main/java/de/baleipzig/eshop/api/dto/ProductDTO.java
@@ -1,11 +1,12 @@
 package de.baleipzig.eshop.api.dto;
 
 import de.baleipzig.eshop.api.foo.ProductType;
+import de.baleipzig.eshop.api.validation.ValueOfEnum;
 
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.Size;
 
-public record ProductDTO(@NotBlank @ValueOfEnum(enumClass = ProductType.class) String productType,
+public record ProductDTO(@NotBlank @ValueOfEnum(enumClass = ProductType.class)String productType,
                          @NotBlank @Size(min = 2, message = "product Name should have at least 2 characters") String name,
                          String property) {
